todo menos mejorRest

+++ RAPPI2/src/UIMain/Restaurante/CambiarPlato.java
@@ -18,7 +18,7 @@ public class CambiarPlato extends OpcionDeMenu {
 		System.out.println("Ingrese la descripción del plato: ");
 		String descripcion = Main.user.next();
 		System.out.println("Ingrese el precio: ");
-		float precio = Main.user.nextFloat();
+		int precio = Main.user.nextInt();
 		System.out.println("ingrese la restriccion de edad del plato: ");
 		int restriccion = Main.user.nextInt();
 		Plato platoCambio = new Plato(nombre, descripcion, precio, restriccion, restaurante);
